requirements.txt
Flask==2.0.1
numpy==1.22.4
Pillow==9.0.1
Werkzeug==2.0.1
gunicorn==20.1.0
poppler-utils